+++ CHANGELOG.md
@@ -45,6 +45,7 @@
     * Fixed tab bars not showing when loading documents containing precomps
     * Fixed crash when opening the most recent file
     * Imported files show their name in the created layer
+    * Fixed Lottie export of transparent color
 
 ## 0.4.3